requirements.txt
APScheduler==3.11.0
nonebot2==2.4.1
nonebot_plugin_alconna==0.54.2
nonebot_plugin_localstore==0.7.3
nonebot_plugin_orm==0.7.6
openai==1.64.0
pydantic==2.10.5
ruamel.yaml==0.18.10
SQLAlchemy==2.0.38
toml>=0.10.2; python_version < '3.11'
websocket_client==1.8.0
watchdog==6.0.0
aiosqlite>=0.17.0
